Templates, Drawing API and more (bugs remain)

+++ src/include/util/types.h
@@ -2,4 +2,8 @@
 
 #include <inttypes.h>
 
-typedef float type;
+typedef double type; 
+/*
+ * Carful with this, there are some places where changing it to float will break things with the OpenGL code. 
+ * For example in draw.cpp: 	"glVertexAttribPointer(0, 2, GL_DOUBLE, GL_FALSE, 2 * sizeof(type), (void *)0);"    where you tell OpenGL that the data is of type GL_DOUBLE. 
+ */
